requirements.txt
Django==2.1.5
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
six==1.12.0
